# apps/backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ElektroPlan",
    "description": "Modulare Plattform fuer Elektrofachbetriebe. Geldbetraege und Mengen werden als Dezimal-String uebertragen (ADR 0005).",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/modules/electrical/floors/{floor_id}/rooms": {
      "get": {
        "tags": [
          "electrical"
        ],
        "summary": "Raeume eines Geschosses",
        "description": "Raeume des Geschosses - nach Raumnummer, dann Name, dann ID sortiert.\n\nKeine Cursor-Pagination: Ein Geschoss hat Raeume in zweistelliger Anzahl.\nEine Seitenmechanik ohne Bedarf waere nur mehr Vertrag zum Pflegen.",
        "operationId": "listElectricalRooms",
        "parameters": [
          {
            "name": "floor_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Floor Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RoomOut"
                  },
                  "title": "Response Listelectricalrooms"
                }
              }
            }
          },
          "404": {
            "description": "Nicht gefunden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "electrical"
        ],
        "summary": "Raum anlegen",
        "description": "Legt einen Raum auf dem Geschoss an. Die Kontur folgt als Waende.",
        "operationId": "createElectricalRoom",
        "parameters": [
          {
            "name": "floor_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Floor Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoomCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoomOut"
                }
              }
            }
          },
          "404": {
            "description": "Nicht gefunden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "409": {
            "description": "Projekt ist archiviert",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "422": {
            "description": "Geometrie oder Eingabe unzulaessig",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/modules/electrical/rooms/{room_id}": {
      "get": {
        "tags": [
          "electrical"
        ],
        "summary": "Raum",
        "description": "Ein Raum samt berechneter Flaeche, Umfang und Konturzustand.",
        "operationId": "getElectricalRoom",
        "parameters": [
          {
            "name": "room_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Room Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoomOut"
                }
              }
            }
          },
          "404": {
            "description": "Nicht gefunden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "electrical"
        ],
        "summary": "Raum bearbeiten",
        "description": "Aendert Name, Raumnummer oder Raumhoehe.",
        "operationId": "updateElectricalRoom",
        "parameters": [
          {
            "name": "room_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Room Id"
            }
          },
          {
            "name": "If-Match",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-Match"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoomUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoomOut"
                }
              }
            }
          },
          "404": {
            "description": "Nicht gefunden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "409": {
            "description": "Versionskonflikt, archiviertes Projekt oder Wand mit Oeffnungen",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "422": {
            "description": "Geometrie oder Eingabe unzulaessig",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "428": {
            "description": "If-Match fehlt",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "electrical"
        ],
        "summary": "Raum loeschen",
        "description": "Entfernt den Raum endgueltig - samt seiner Waende und Oeffnungen.",
        "operationId": "deleteElectricalRoom",
        "parameters": [
          {
            "name": "room_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Room Id"
            }
          },
          {
            "name": "If-Match",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-Match"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "description": "Nicht gefunden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "409": {
            "description": "Versionskonflikt, archiviertes Projekt oder Wand mit Oeffnungen",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "422": {
            "description": "Geometrie oder Eingabe unzulaessig",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "428": {
            "description": "If-Match fehlt",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RoomOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "floor_id": {
            "type": "string",
            "format": "uuid",
            "title": "Floor Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "room_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Room Number"
          },
          "height_mm": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Height Mm"
          },
          "effective_height_mm": {
            "type": "integer",
            "title": "Effective Height Mm"
          },
          "contour_status": {
            "type": "string",
            "enum": [
              "draft",
              "valid"
            ],
            "title": "Contour Status"
          },
          "wall_count": {
            "type": "integer",
            "title": "Wall Count"
          },
          "area_mm2": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Area Mm2"
          },
          "area_m2": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Area M2"
          },
          "perimeter_mm": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Perimeter Mm"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "floor_id",
          "name",
          "room_number",
          "height_mm",
          "effective_height_mm",
          "contour_status",
          "wall_count",
          "area_mm2",
          "area_m2",
          "perimeter_mm",
          "version",
          "created_at",
          "updated_at"
        ],
        "title": "RoomOut",
        "description": "Ein Raum samt abgeleiteten Konturwerten.\n\nFlaeche, Umfang, Wandzahl und Konturzustand sind **berechnet** und nicht\ngespeichert (ADR 0013). Sie fehlen (``null``), solange die Kontur nicht\ngeschlossen ist."
      },
...
    }
  }
}